# HiСlass.Backend/HiClass.API/openapi.yml
openapi: 3.0.1
info:
  title: HiClass API
  description: ASP.NET Core Web API for HiClass
  version: v1
paths:
  /api/Class/create-class:
    post:
      tags:
        - Class
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - DisciplineTitles
                - GradeNumber
                - LanguageTitles
                - Title
              type: object
              properties:
                Title:
                  type: string
                GradeNumber:
                  type: integer
                  format: int32
                LanguageTitles:
                  type: array
                  items:
                    type: string
                DisciplineTitles:
                  type: array
                  items:
                    type: string
            encoding:
              Title:
                style: form
              GradeNumber:
                style: form
              LanguageTitles:
                style: form
              DisciplineTitles:
                style: form
      responses:
        '200':
          description: OK
  '/api/Class/class-profile/{classId}':
    get:
      tags:
        - Class
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
  '/api/Class/edit-class/{classId}':
    put:
      tags:
        - Class
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditClassRequestDto'
          text/json:
            schema:
              $ref: '#/components/schemas/EditClassRequestDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/EditClassRequestDto'
      responses:
        '200':
          description: OK
  '/api/Class/delete-class/{classId}':
    delete:
      tags:
        - Class
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
  /api/Image/set-user-image:
    put:
      tags:
        - Image
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - ImageFormFile
              type: object
              properties:
                ImageFormFile:
                  type: string
                  format: binary
            encoding:
              ImageFormFile:
                style: form
      responses:
        '200':
          description: OK
  /api/Image/edit-user-image:
    put:
      tags:
        - Image
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - ImageFormFile
              type: object
              properties:
                ImageFormFile:
                  type: string
                  format: binary
            encoding:
              ImageFormFile:
                style: form
      responses:
        '200':
          description: OK
  /api/Image/set-user-banner-image:
    put:
      tags:
        - Image
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - ImageFormFile
              type: object
              properties:
                ImageFormFile:
                  type: string
                  format: binary
            encoding:
              ImageFormFile:
                style: form
      responses:
        '200':
          description: OK
  /api/Image/edit-user-banner-image:
    put:
      tags:
        - Image
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - ImageFormFile
              type: object
              properties:
                ImageFormFile:
                  type: string
                  format: binary
            encoding:
              ImageFormFile:
                style: form
      responses:
        '200':
          description: OK
  '/api/Image/set-class-image/{classId}':
    put:
      tags:
        - Image
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - ImageFormFile
              type: object
              properties:
                ImageFormFile:
                  type: string
                  format: binary
            encoding:
              ImageFormFile:
                style: form
      responses:
        '200':
          description: OK
  '/api/Image/edit-class-image/{classId}':
    put:
      tags:
        - Image
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - ImageFormFile
              type: object
              properties:
                ImageFormFile:
                  type: string
                  format: binary
            encoding:
              ImageFormFile:
                style: form
      responses:
        '200':
          description: OK
...
  /api/User/register:
    post:
      tags:
        - User
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequestDto'
          text/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequestDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequestDto'
      responses:
        '200':
          description: OK
  /api/User/login:
    post:
      tags:
        - User
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequestDto'
          text/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequestDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UserLoginRequestDto'
      responses:
        '200':
          description: OK
  /api/User/verify-email:
    post:
      tags:
        - User
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailVerificationRequestDto'
          text/json:
            schema:
              $ref: '#/components/schemas/EmailVerificationRequestDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/EmailVerificationRequestDto'
      responses:
        '200':
          description: OK
  /api/User/reverify-email:
    post:
      tags:
        - User
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailReVerificationRequestDto'
          text/json:
            schema:
              $ref: '#/components/schemas/EmailReVerificationRequestDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/EmailReVerificationRequestDto'
      responses:
        '200':
          description: OK
  /api/User/create-account:
    put:
      tags:
        - User
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - CityLocation
                - CountryLocation
                - DisciplineTitles
                - FirstName
                - GradesEnumerable
                - IsAnExpert
                - IsATeacher
                - LanguageTitles
                - LastName
              type: object
              properties:
                FirstName:
                  type: string
                LastName:
                  type: string
                IsATeacher:
                  type: boolean
                IsAnExpert:
                  type: boolean
                CityLocation:
                  type: string
                CountryLocation:
                  type: string
                InstitutionDto.Types:
                  type: array
                  items:
                    type: string
                InstitutionDto.Address:
                  type: string
                InstitutionDto.Title:
                  type: string
                DisciplineTitles:
                  type: array
                  items:
                    type: string
                LanguageTitles:
                  type: array
                  items:
                    type: string
                GradesEnumerable:
                  type: array
                  items:
                    type: integer
                    format: int32
            encoding:
              FirstName:
                style: form
              LastName:
                style: form
              IsATeacher:
                style: form
              IsAnExpert:
                style: form
              CityLocation:
                style: form
              CountryLocation:
                style: form
              InstitutionDto.Types:
                style: form
              InstitutionDto.Address:
                style: form
              InstitutionDto.Title:
                style: form
              DisciplineTitles:
                style: form
              LanguageTitles:
                style: form
              GradesEnumerable:
                style: form
      responses:
        '200':
          description: OK
  /api/User/forgot-password:
    post:
      tags:
        - User
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequestDto'
          text/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequestDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequestDto'
      responses:
        '200':
          description: OK
  /api/User/check-reset-password-code:
    post:
      tags:
        - User
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckResetPasswordCodeDto'
          text/json:
            schema:
              $ref: '#/components/schemas/CheckResetPasswordCodeDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CheckResetPasswordCodeDto'
      responses:
        '200':
          description: OK
  /api/User/reset-password:
    post:
      tags:
        - User
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequestDto'
          text/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequestDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequestDto'
      responses:
        '200':
          description: OK
...
  '/api/User/other-userprofile/{userId}':
    get:
      tags:
        - User
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
components:
  schemas:
...
    CheckResetPasswordCodeDto:
      required:
        - resetCode
      type: object
      properties:
        resetCode:
          minLength: 1
          type: string
      additionalProperties: false
...
    EditClassRequestDto:
      required:
        - disciplineTitles
        - gradeNumber
        - languageTitles
        - title
      type: object
      properties:
        title:
          minLength: 1
          type: string
        gradeNumber:
          type: integer
          format: int32
        languageTitles:
          type: array
          items:
            type: string
        disciplineTitles:
          type: array
          items:
            type: string
      additionalProperties: false
...
    EmailReVerificationRequestDto:
      required:
        - email
      type: object
      properties:
        email:
          minLength: 1
          type: string
      additionalProperties: false
    EmailVerificationRequestDto:
      required:
        - email
        - verificationCode
      type: object
      properties:
        email:
          minLength: 1
          type: string
        verificationCode:
          minLength: 1
          type: string
      additionalProperties: false
...
    ForgotPasswordRequestDto:
      required:
        - email
      type: object
      properties:
        email:
          minLength: 1
          type: string
      additionalProperties: false
...
    ResetPasswordRequestDto:
      required:
        - password
      type: object
      properties:
        password:
          minLength: 1
          type: string
      additionalProperties: false
...
    UserLoginRequestDto:
      required:
        - email
        - password
      type: object
      properties:
        email:
          minLength: 1
          type: string
          format: email
        password:
          minLength: 1
          type: string
      additionalProperties: false
...
    UserRegisterRequestDto:
      required:
        - email
        - password
      type: object
      properties:
        email:
          minLength: 1
          type: string
          format: email
        password:
          minLength: 1
          type: string
      additionalProperties: false
  securitySchemes:
    Bearer:
      type: http
      description: JWT Authorization header using the Bearer scheme.
      scheme: Bearer
      bearerFormat: JWT
security:
  - Bearer: [ ]